Add files via upload
These files display the entire process for creating an interactive game console that is solely functional from the IDE's terminal. The game console is composed of three game: MyWordle, Rock Paper Scissors (vs. AI or another player), and Aggie Trivia. There is a main menu that is accessible throughout the entirety of the gameplay cycle. This was my fall semester freshman final project (fall 2022).